+++ algorithm/shacrypt/doc.go
@@ -1,5 +1,5 @@
-// Package sha2crypt provides helpful abstractions for an implementation of SHA2Crypt and implements
+// Package shacrypt provides helpful abstractions for an implementation of SHA2Crypt and implements
 // github.com/go-crypt/crypt interfaces.
 //
 // This implementation is loaded by the encoding.NewDefaultDecoder and encoding.NewDecoderAll.
-package sha2crypt
+package shacrypt
