+++ go.mod
@@ -2,6 +2,6 @@ module github.com/go-crypt/crypt
 
 go 1.19
 
-require github.com/go-crypt/x v0.1.5
+require github.com/go-crypt/x v0.1.6
 
 require golang.org/x/sys v0.2.0 // indirect
